adds layout container
